separated Dom methods
