+++ packages/dom/src/blockElements.ts
@@ -1,0 +1,47 @@
+/**
+ * Return array of names of block html elements
+ * @returns
+ */
+export function blockElements(): string[] {
+  return [
+    'address',
+    'article',
+    'aside',
+    'blockquote',
+    'canvas',
+    'div',
+    'dl',
+    'dt',
+    'fieldset',
+    'figcaption',
+    'figure',
+    'footer',
+    'form',
+    'h1',
+    'h2',
+    'h3',
+    'h4',
+    'h5',
+    'h6',
+    'header',
+    'hgroup',
+    'hr',
+    'li',
+    'main',
+    'nav',
+    'noscript',
+    'ol',
+    'output',
+    'p',
+    'pre',
+    'ruby',
+    'section',
+    'table',
+    'tbody',
+    'thead',
+    'tr',
+    'tfoot',
+    'ul',
+    'video',
+  ];
+}
